requirements.txt
selenium>=4.15.2
pandas>=2.1.0
webdriver-manager>=4.0.1
openpyxl>=3.1.2
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-api-python-client==2.86.0 